requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
plotly==5.17.0
groq==0.4.1
gspread==6.0.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0